refactor(fitness): split fitness_tools.py into per-concern modules
fitness_tools.py had grown to 1177 lines holding the database schema, the
Arbox HTTP client, the WOD-programming parser, and all 13 tools at once —
the point where finding the code for a change meant scrolling past four
unrelated concerns, and where adding the Arbox registration write path
would have made it worse.

Split along the layout tools/travel/ already uses: a private _db.py for the
schema, connection, migrations and the helpers every tool shares; a private
_arbox.py for the gym's API surface and track parsing; then one module per
group of tools — classes.py (Arbox reads), logs.py (lifts, runs, WOD
results), reports.py (weekly summary, adherence), plans.py, query.py.
Largest module is now 295 lines. __init__.py imports each tool explicitly,
mirroring tools/travel/__init__.py, so the @tool_register side-effects still
run on package import.

No behavior change, and that is checked rather than asserted: every
top-level segment was relocated by AST extraction rather than retyped, and
all 33 compare AST-identical against their previous definitions. The
registered tool surface — 13 names, namespaces, destructive flags, scopes,
docstrings, arg schemas — diffs clean before and after. A symtable pass over
each new module caught three imports that the split would otherwise have
dropped (timedelta/timezone in plans, timezone in reports, re in query);
those fail only at call time, so a plain import check would have missed them.

gateway/apps/fitness.py takes _FITNESS_RO_URI, _fmt_pace, _target_for_week
and ISRAEL_TZ from tools.fitness._db now. Importing that module still runs
_init_db(), so the app surface's guarantee that the schema exists on an
instance that has never used the skill in chat is preserved.
